+++ raillo-common/build.gradle
@@ -2,6 +2,11 @@
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+    // CommonExceptionHandler에서 AccessDeniedException(도메인 무관 403)을 참조.
+    // implementation으로 두면 인증 미사용 서비스에도 시큐리티 기본 보안이 강제되므로 compileOnly.
+    compileOnly 'org.springframework.boot:spring-boot-starter-security'
 }
 
 jar {
