+++ raillo-common/src/main/java/com/sudo/raillo/common/exception/CommonExceptionHandler.java
@@ -1,16 +1,15 @@
-package com.sudo.raillo.global.exception;
+package com.sudo.raillo.common.exception;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.springframework.dao.InvalidDataAccessApiUsageException;
-import org.springframework.data.redis.RedisConnectionFailureException;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.MissingServletRequestParameterException;
@@ -18,38 +17,23 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 
-import com.sudo.raillo.auth.exception.TokenError;
-import com.sudo.raillo.common.exception.BusinessException;
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.common.exception.ExternalApiException;
-import com.sudo.raillo.common.exception.GlobalError;
-import com.sudo.raillo.global.redis.exception.RedisError;
-import com.sudo.raillo.global.redis.exception.RedisException;
+import com.sudo.raillo.common.response.ErrorResponse;
 
-import io.jsonwebtoken.io.SerializationException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
+@Order(Ordered.LOWEST_PRECEDENCE)
 @RestControllerAdvice
-public class GlobalExceptionHandler {
+public class CommonExceptionHandler {
 
-	/**
-	 * 접근 권한 예외 처리
-	 */
 	@ExceptionHandler(AccessDeniedException.class)
 	public ResponseEntity<ErrorResponse> handleAccessDeniedException(AccessDeniedException ex) {
 		ErrorResponse errorResponse = ErrorResponse.of(GlobalError.FORBIDDEN_ACCESS, ex.getMessage());
-
-		return ResponseEntity
-			.status(HttpStatus.FORBIDDEN)
-			.body(errorResponse);
+		return ResponseEntity.status(HttpStatus.FORBIDDEN).body(errorResponse);
 	}
 
-	/**
-	 * @RequestBody 누락 처리 : HttpMessageNotReadableException
-	 */
 	@ExceptionHandler(HttpMessageNotReadableException.class)
 	public ResponseEntity<ErrorResponse> handleHttpMessageNotReadable(
 		HttpMessageNotReadableException e, HttpServletRequest request) {
@@ -67,9 +51,6 @@ public class GlobalExceptionHandler {
 		);
 	}
 
-	/**
-	 * @RequestBody 유효성 검사 실패 처리 : MethodArgumentNotValidException
-	 */
 	@ExceptionHandler(MethodArgumentNotValidException.class)
 	public ResponseEntity<ErrorResponse> handleMethodArgumentNotValid(MethodArgumentNotValidException ex) {
 		List<FieldError> fieldErrors = ex.getBindingResult().getFieldErrors();
@@ -79,9 +60,6 @@ public class GlobalExceptionHandler {
 		return ResponseEntity.badRequest().body(errorResponse);
 	}
 
-	/**
-	 * @RequestParam 누락 처리 : MissingServletRequestParameterException
-	 */
 	@ExceptionHandler(MissingServletRequestParameterException.class)
 	public ResponseEntity<ErrorResponse> handleMissingServletRequestParameter(
 		MissingServletRequestParameterException ex) {
@@ -92,9 +70,6 @@ public class GlobalExceptionHandler {
 		return ResponseEntity.badRequest().body(errorResponse);
 	}
 
-	/**
-	 * @PathVariable, @RequestParam 유효성 검사 실패 처리 : ConstraintViolationException
-	 */
 	@ExceptionHandler(ConstraintViolationException.class)
 	public ResponseEntity<ErrorResponse> handleConstraintViolation(ConstraintViolationException ex) {
 		Map<String, String> errors = new HashMap<>();
@@ -109,9 +84,6 @@ public class GlobalExceptionHandler {
 		return ResponseEntity.badRequest().body(errorResponse);
 	}
 
-	/**
-	 * 비즈니스 예외 처리
-	 */
 	@ExceptionHandler(BusinessException.class)
 	public ResponseEntity<ErrorResponse> handleBusinessException(BusinessException ex) {
 		ErrorResponse errorResponse = ErrorResponse.of(ex.getErrorCode());
@@ -120,11 +92,8 @@ public class GlobalExceptionHandler {
 		return ResponseEntity.status(ex.getErrorCode().getStatus()).body(errorResponse);
 	}
 
-	/**
-	 * 외부 API  예외 처리
-	 */
 	@ExceptionHandler(ExternalApiException.class)
-	public ResponseEntity<ErrorResponse> handleBusinessException(ExternalApiException ex) {
+	public ResponseEntity<ErrorResponse> handleExternalApiException(ExternalApiException ex) {
 		Map<String, Object> details = Map.of(
 			"provider", ex.getProvider(),
 			"type", ex.getErrorType()
@@ -136,9 +105,13 @@ public class GlobalExceptionHandler {
 		return ResponseEntity.status(ex.getHttpStatus()).body(response);
 	}
 
-	/**
-	 * 예상하지 못한 모든 예외 처리 : Exception
-	 */
+	@ExceptionHandler(IllegalArgumentException.class)
+	public ResponseEntity<ErrorResponse> handleIllegalArgumentException(IllegalArgumentException ex) {
+		ErrorResponse errorResponse = ErrorResponse.of(GlobalError.INVALID_REQUEST_PARAM);
+		log.warn("Invalid argument: {}", ex.getMessage());
+		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
+	}
+
 	@ExceptionHandler(Exception.class)
 	public ResponseEntity<ErrorResponse> handleException(Exception ex, WebRequest request) {
 		ErrorResponse errorResponse = ErrorResponse.of(
@@ -150,69 +123,6 @@ public class GlobalExceptionHandler {
 		return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
 	}
 
-	/**
-	 * 비밀번호 불일치 예외 처리
-	 */
-	@ExceptionHandler(BadCredentialsException.class)
-	public ResponseEntity<ErrorResponse> handleBadCredentialsException(BadCredentialsException ex) {
-		ErrorResponse errorResponse = ErrorResponse.of(TokenError.INVALID_PASSWORD);
-		log.warn("Bad credentials: {}", ex.getMessage());
-		return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(errorResponse);
-	}
-
-	/**
-	 * Redis: 연결 실패 예외 처리
-	 * */
-	@ExceptionHandler(RedisConnectionFailureException.class)
-	public ResponseEntity<ErrorResponse> handleRedisConnectionFailure(RedisConnectionFailureException ex) {
-		ErrorResponse errorResponse = ErrorResponse.of(RedisError.REDIS_CONNECT_FAIL);
-		log.warn("Redis connection failure: {}", ex.getMessage());
-		return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-	}
-
-	/**
-	 * Redis: 직렬화 및 역직렬화 예외 처리
-	 * */
-	@ExceptionHandler(SerializationException.class)
-	public ResponseEntity<ErrorResponse> handleRedisSerializationException(Exception ex) {
-		ErrorResponse errorResponse = ErrorResponse.of(RedisError.SERIALIZATION_FAIL);
-		log.warn("Redis serialization failure: {}", ex.getMessage());
-		return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
-	}
-
-	/**
-	 * Redis: 잘못된 API 사용 예외 처리
-	 * */
-	@ExceptionHandler(InvalidDataAccessApiUsageException.class)
-	public ResponseEntity<ErrorResponse> handleRedisApiUsageException(InvalidDataAccessApiUsageException ex) {
-		ErrorResponse errorResponse = ErrorResponse.of(RedisError.INVALID_DATA_ACCESS);
-		log.warn("Redis API usage failure: {}", ex.getMessage());
-		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
-	}
-
-	/**
-	 * 예상하지 못한 Redis의 예외 처리
-	 * */
-	@ExceptionHandler(RedisException.class)
-	public ResponseEntity<ErrorResponse> handleRedisException(RedisException ex) {
-		ErrorResponse errorResponse = ErrorResponse.of(ex.getErrorCode());
-		log.warn("Redis exception occurred: {}", ex.getMessage());
-		return ResponseEntity.status(ex.getErrorCode().getStatus()).body(errorResponse);
-	}
-
-	/**
-	 * 잘못된 인수 전달 시 예외 처리
-	 * */
-	@ExceptionHandler(IllegalArgumentException.class)
-	public ResponseEntity<ErrorResponse> handleIllegalArgumentException(IllegalArgumentException ex) {
-		ErrorResponse errorResponse = ErrorResponse.of(GlobalError.INVALID_REQUEST_PARAM);
-		log.warn("Invalid argument: {}", ex.getMessage());
-		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
-	}
-
-	/**
-	 * 비즈니스 예외 로깅
-	 */
 	private void logBusinessException(BusinessException ex) {
 		if (ex.getErrorCode().getStatus().is5xxServerError()) {
 			log.error("Business exception occurred", ex);
@@ -221,9 +131,6 @@ public class GlobalExceptionHandler {
 		}
 	}
 
-	/**
-	 * 외부 API 예외 로깅
-	 */
 	private void logExternalApiException(ExternalApiException ex) {
 		log.warn("External API Error | Provider={} | ErrorType={} | HTTP={} | Code={} | Message={}",
 			ex.getProvider(),
