+++ raillo-common/src/main/java/com/sudo/raillo/common/response/GlobalResponseHandler.java
@@ -1,4 +1,4 @@
-package com.sudo.raillo.global.success;
+package com.sudo.raillo.common.response;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.http.MediaType;
