+++ raillo-core/src/main/java/com/sudo/raillo/auth/docs/AuthControllerDoc.java
@@ -5,8 +5,8 @@ import com.sudo.raillo.auth.application.dto.request.SignUpRequest;
 import com.sudo.raillo.auth.application.dto.response.LoginResponse;
 import com.sudo.raillo.auth.application.dto.response.ReissueTokenResponse;
 import com.sudo.raillo.auth.application.dto.response.SignUpResponse;
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.ErrorResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
