+++ raillo-core/src/main/java/com/sudo/raillo/auth/presentation/EmailAuthController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 import com.sudo.raillo.auth.application.EmailAuthService;
 import com.sudo.raillo.member.application.MemberService;
 import com.sudo.raillo.auth.application.dto.request.SendCodeRequest;
