+++ raillo-core/src/main/java/com/sudo/raillo/auth/success/AuthSuccess.java
@@ -2,7 +2,7 @@ package com.sudo.raillo.auth.success;
 
 import org.springframework.http.HttpStatus;
 
-import com.sudo.raillo.global.success.SuccessCode;
+import com.sudo.raillo.common.response.SuccessCode;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
