+++ raillo-core/src/main/java/com/sudo/raillo/booking/docs/PendingBookingControllerDoc.java
@@ -8,8 +8,8 @@ import com.sudo.raillo.booking.application.dto.request.PendingBookingCreateReque
 import com.sudo.raillo.booking.application.dto.request.PendingBookingDeleteRequest;
 import com.sudo.raillo.booking.application.dto.response.PendingBookingCreateResponse;
 import com.sudo.raillo.booking.application.dto.response.PendingBookingDetailResponse;
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.ErrorResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
