+++ raillo-core/src/main/java/com/sudo/raillo/booking/success/BookingSuccess.java
@@ -3,7 +3,7 @@ package com.sudo.raillo.booking.success;
 import org.springframework.boot.autoconfigure.graphql.GraphQlProperties;
 import org.springframework.http.HttpStatus;
 
-import com.sudo.raillo.global.success.SuccessCode;
+import com.sudo.raillo.common.response.SuccessCode;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
