+++ raillo-core/src/main/java/com/sudo/raillo/member/docs/MemberControllerDoc.java
@@ -1,7 +1,7 @@
 package com.sudo.raillo.member.docs;
 
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.ErrorResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 import com.sudo.raillo.member.application.dto.response.MemberInfoResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
