+++ raillo-core/src/main/java/com/sudo/raillo/member/docs/MemberFindControllerDoc.java
@@ -3,8 +3,8 @@ package com.sudo.raillo.member.docs;
 import com.sudo.raillo.auth.application.dto.request.VerifyCodeRequest;
 import com.sudo.raillo.auth.application.dto.response.SendCodeResponse;
 import com.sudo.raillo.auth.application.dto.response.TemporaryTokenResponse;
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.ErrorResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 import com.sudo.raillo.member.application.dto.request.FindMemberNoRequest;
 import com.sudo.raillo.member.application.dto.request.FindPasswordRequest;
 import com.sudo.raillo.member.application.dto.response.VerifyMemberNoResponse;
