+++ raillo-core/src/main/java/com/sudo/raillo/member/docs/MemberUpdateControllerDoc.java
@@ -1,7 +1,7 @@
 package com.sudo.raillo.member.docs;
 
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.ErrorResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 import com.sudo.raillo.auth.application.dto.request.SendCodeRequest;
 import com.sudo.raillo.member.application.dto.request.UpdateEmailRequest;
 import com.sudo.raillo.member.application.dto.request.UpdatePasswordRequest;
