+++ raillo-core/src/main/java/com/sudo/raillo/train/docs/CachedTrainSearchControllerDoc.java
@@ -1,7 +1,7 @@
 package com.sudo.raillo.train.docs;
 
-import com.sudo.raillo.global.exception.error.ErrorResponse;
-import com.sudo.raillo.global.success.SuccessResponse;
+import com.sudo.raillo.common.response.ErrorResponse;
+import com.sudo.raillo.common.response.SuccessResponse;
 import com.sudo.raillo.train.application.dto.response.OperationCalendarItemResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
